Arranged all the java files in respective Folders
