+++ Java/Others/NQueenProblem.java
